requirements.txt
alembic==1.16.2
anyio==4.9.0
attrs==25.4.0
babel==2.18.0
backoff==2.2.1
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
decorator==5.2.1
dill==0.4.1
distlib==0.4.0
exceptiongroup==1.3.1
filelock==3.24.2
greenlet==3.2.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
iniconfig==2.1.0
llvmlite==0.46.0
Mako==1.3.10
MarkupSafe==3.0.2
movielite==0.2.2
multiprocess==0.70.19
numba==0.64.0
numpy==2.2.4
opencv-python==4.13.0.92
packaging==24.2
pictex==2.1.0
pillow==10.4.0
pipenv==2026.0.3
platformdirs==4.9.2
pluggy==1.5.0
proglog==0.1.10
pyaes==1.6.1
pybind11==3.0.2
Pygments==2.19.2
Pyrogram==2.0.106
PySocks==1.7.1
pytest==8.4.2
pytest-asyncio==0.26.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==22.0
regex==2026.2.28
requests==2.32.5
six==1.17.0
skia-python==144.0.post1
sniffio==1.3.1
SQLAlchemy==2.0.41
stretchable==1.1.8
tinytag==2.2.0
tomli==2.4.0
tqdm==4.67.1
typing_extensions==4.15.0
uharfbuzz==0.53.3
urllib3==2.3.0
virtualenv==20.36.1
yt-dlp==2025.6.30
